Font Size Correction
The date and temperature font sizes have been scaled back to fit the screen better. 

The loading text for the weather was dropping onto a second line. This has been replaced with "...".

After looking over other watchfaces, and options for the broken clouds condition icon, I'll be omitting it and leaving the cloudy partially cloudy icon in it's place.